requirements.txt
slack-sdk
slack-bolt
Flask
python-dotenv
langchain
openai